chore(nimbus): split v5 api serializer tests
Because

* There are too many API V5 serializer tests in one file
* Makes it hard to navigate
* Turns out we also had a bunch of duplicated tests and nobody noticed

This commit

* Splits that big file up into separate test files
